Add spheres to core functions
Spheres will output a series of concentric spheres, as a triangle graph
between 0 and 1, at the specified frequency. It has 2D, 3D and 4D
versions.

The cylinders generator was updated to call the spheres functions. It
always calls the 2D version, because the 2D version is equivalent to
what cylinders is already built to do. Spheres is a generalization of
that to higher dimensions.

Also fixed the cylinders example.