Move fbm to core
Add new fbm functions to core, and update the NoiseFn implementation to
use it. The new fbm functions can work with any noise function, unlike
the old ones that were hardcoded to use perlin noise. It also passes the
current octave into the noise function, allowing the caller to vary the
function based on the octive by whatever means it wants, such as adding
an offset or variant noise function to avoid repetition near the origin.

There is also a new example which demonstrates 2d, 3d and 4d fbm with
both perlin and open simplex noise.