Add perlin variant functions
The variant functions take an extra argument to essentially specify a
sub-seed. It still uses the same permutation table, but hashes against
the variant seed as well. This is useful when doing things like FBM,
where you're taking many instances of the same noise function and want
each one to be different.

Also added fbm_perlin_2d, fbm_perlin_3d and fbm_perlin_4d as shortcut
functions that do fbm on perlin noise with the variants.