Break out different versions of simplex functions
The 2D, 3D and 4D versions of simplex have now each have a deriv version
that also returns the derivative, a variant version that allows for a
sub-seed, and a variant_deriv version that does both.

The benches are updated to bench the base and deriv versions separately.
Previously, the deriv version was the only version. The benches show
that the base version is significantly faster, which justifies the
separation, as when you don't need the derivative you can gain
substantial performance by calling the other version.

The difference between the deriv and base version is merely that the
base version discards the derivative from the return. However, because
it's all inlined, this provides the compiler with the opportunity to
strip out all the unused calculations.

Also updated the fbm module to have simplex shortcut fbm functions,
using the new variant versions.