Unify the internal vector type into a macro
Previously had a general trait for vectors of any number of dimensions,
and then had a seperate implemention for 2d, 3d and 4d vectors, which
were almost entirely redundant.

Now, there's a big vector_type macro, which takes the name of each
dimension and builds a vector type with that number of dimensions. The
old Vector trait was removed, as nothing used it directly.

This should reduce the opportunities for bugs, as any change to one type
of vector is automatically replicated across all of them.