Simplify PlaneMapBuilder a little
It no longer passes through the PermutationTable. Instead, the function
it's given only takes a point. This means it can be used with functions
that don't need a permutation table.

Also take advantage of this to update the checkerboard example to use
the core function directly instead of the NoiseFn struct.