Vectors can now add or subtract scalars
Adding or subtracting a scalar to a vector adds or subtracts that value
to all elements of the vector. In this way, it is equivalent to
multiplying or dividing by a scalar.

Sadly this made some type inference hiccups, so this also required
putting explicit types in a number of places where they hadn't been
needed before.