Update simplex to use vectors
The simplex core functions now take a Vector2/3/4 as it's input, instead
of taking an array. The 2D function is also converted to using entirely
vectors for all it's internal logic, but for some reason doing the same
for the 3D and 4D functions caused a significant performance regression,
so they'll stay the way they are for now, just converting out of the
vector at the start, and converting back at the end.