Update super simplex to use vectors
The super simplex core functions now take a Vector2/3 as it's input,
instead of taking an array. But the 2D function is converted from using
vectors for all it's internal logic, because breaking it out and doing
the logic directly on the component scalars got a significant
improvement on the benchmarks. This did not help for the 3D function, so
that was left using the vectors.