Remove extraneous distance functions from worley
Worley noise had 5 distance functions. Of them, Euclidean and Euclidean
Square are the only ones of which I am convinced of the value. The
others (Manhattan, Chebyshev and Quadratic) can be re-added easily
enough if concrete use cases come up.

Also, I'm fairly certain Quadratic was completely broken. The output
looked very wrong, but as I have no idea what it's supposed to do or
look like, I had no basis to fix it.

The main reason for stripping them out is that each distance function
that's being maintained needs to be benchmarked, but that meant that
with 5 distance functions by 3 dimensions, with single and grid variants
there were 40 worley benchmarks to run, taking a minimum of 5:20 minutes
to run. Which is really bad for iteration. Stripping it down to just the
most useful ones, instead of going for completeness, makes it much
easier to take the time to maximize the performance and flexibility of
the worley noise functions.