Make worley use lookup tables for seed points
Previously, worley noise hashed the seed point out of the index value.
Now, there's just a gigantic const lookup table for each of 2d, 3d and
4d. This is both MUCH more performant, and the distribution of points in
these lookup tables guarantees no discontinuities, which was not true of
the old method. Further, these distributions favor points near the
extremes over points in the middle of the cell, which tends to create a
more random looking noise. The old distribution favored points in the
center, which created a more gridlike noise.

The new worley_table_builder example is not a real example, but is just
a convenient way to package up a related utility. This is a small
program that constructs the worley lookup tables.