Simplify and speed up worley noise
The worley noise functions no longer take a distance function, or a
return type. Instead, the base functions always return the distance
squared, the cell and the point. Then for each of 2D, 3D and 4D, there
is a _value functon, a _range function and a _range_sqr function, which
calls through to this base function, and utilizes the relevant part of
the result.

Removing the distance functions seemed justified given that the only
distance functions that remained were euclidean and euclidean squared,
and those will always return the same point, and the first could be
easily derived from the second. So now it does all calculations with
euclidean squared, and takes the square root at the end if needed.

Converting all intermediate range comparisons to range squared resulted
in a significant speedup for all the non-squared worley functions.