Add 2-point worley noise
Also do a bunch of fixup in the worley noise in general.
- Extract get_point_Xd functions from the worley functions, so they can
be used in both the 1-point and 2-point variants.
- Rename base_worley_Xd to base_worley_Xd_1point.
- Add base_worley_Xd_2point functions that return the nearest 2 points.
- Change the lookup tables to return seed points in the range of 0-1
instead of -0.5 to 0.5. The 1-point functions were doing some rather
unintuitive math that necessitated that range, but switching it to work
with the 0-1 range lost nothing, and the 2-point function is also best
with the 0-1 range.
- The test_point! macros are all updated to also take the list of ranges
to check against, so all those if statements can become part of the
macro. This makes the functions more readable.
- Add sqr function to vectors, because it means the cell_range
calculations no longer need to call powf(2.0).
- Add n_minus function to vectors, as it eliminated some calls to map,
which made the expressions more readable.
- Add the worley_Xd_border functions for the F2-F1 worley variants with
the 2-point functions.
- Add examples and benchmarks with the new border functions.