Replace floor on vectors with floor_to_isize
Nearly every noise function was calling floor, and this compiled down to
a slow function call, one for each dimension. The new floor_to_isize
function doesn't call the standard floor function, but instead casts the
value to an isize, which implicitly truncates. This is much faster. For
all the noise functions that need the floored value as a f64 still, they
now convert back to f64 after calling floor_to_isize.

This got a significant speed improvement on nearly every noise function.