Add weighted and ridged fbm fractals
The existing core::fbm module was refactored to use macros to cut down
on the boilerplate. A new fbm_weighted_Xd set of functions was added, in
order to support multifractals that have an additional weight mechanism
beyond just the normal lacunarity. The fbm_ridged_Xd functions implement
a particular weighting mechanism on top of that, based on the existing
RidgedMulti noise function, but generalized such that any base noise
function could be plugged into it. There are convenience functions added
for combining fbm_ridged with perlin, perlin_surflet, open_simplex, and
simplex noise.

The fbm example was also simplified by having the code to output an
image factored into it's own function, and a new fbm_ridged example was
added which follows the same pattern, but with the new fbm_ridged noise.