Add the select! macro for blending between values
This macro allows you to blend between any number of values based on a
single control value. Importantly, the values to blend between are given
as blocks, and it will always execute only one or two of those blocks,
being the ones that are needed for the current control value.

The syntax looks like:

select!(control,
	{
		value1
	} => blend_min1, blend_max1 => {
		value2
	} => blend_min2, blend_max2 => {
		value3
	}
)

Where blend_min and blend_max represent the range of values where it'll
be blending between the previous value and the next value. If the
control is between the previous blend_max and the next blend_min, then
it will just use that value and not do any blending.