Remove NoiseFns
This completes the conversion from using the NoiseFn trait back to just
simple free functions and closures. This makes for a much simpler, more
flexible and easier to use API.

In doing this, math::interpolate::cubic and math::scale_shift became
unused, so they were also removed. PermutationTable had a couple unit
tests which referenced the NoiseFn trait, but all they did was confirm
that the PermutationTable didn't crash when used, which makes them very
low value tests. So they were also deleted.

The NoiseMapBuilder needed some refactoring to no longer use the NoiseFn
trait. Also, the CylinderMapBuilder was deleted, as nothing was using
it, and it's not really the job of a noise library to come with
comprehensive tools for building images. These tools will soon be
stripped down to only what is needed for the examples, and then moved
entirely behind a feature flag, to avoid bloating the library. If people
need this functionality, it should come from a different crate.