Image support is now just what's need by examples
Previously, there was a full-featured image creation library here, with
the ability to make planar, spherical or cylinderical maps, do lighting
and make them seamless. Which was all very fancy, but honestly out of
scope for a noise library. If a consumer needs this functionality, say
to do procedurally generated textures, it should be provided by another
crate.

It's now been stripped down to JUST the functionality needed by the
examples, and all of it is entirely hidden behind the "images" feature,
so it won't bloat consumers of this library.