Fix curve and scale of simplex 2D
2D simplex was using a surflet curve of x^4, in the range of 0-0.5,
which would not be smooth because the curve isn't flat at both ends.
Also, it didn't output a range from 0-1, and thus required a scaling
factor at the end, and didn't have the right scaling factor.

It now uses the curve 2x^2+x^4, in the range of 0-1. This is flat at
both ends, and it outputs a range from 0-1, so no scaling factor is
needed.

Also renamed the distance functions to corner_offset, because they were
actually storing a 2D vector. Distance is not a vector, but rather the
magnitude of a vector.

This does slightly reduce the performance, because it has to do a couple
extra math operations. But that's the price of correctness.